Fixes a 64 bit warning - fixes the image viewer in 64bit mode.

+++ SpotSeeker/ViewControllers/SpaceImageScrollPageViewController.m
@@ -9,7 +9,7 @@
 #import "SpaceImageScrollPageViewController.h"
 
 @interface SpaceImageScrollPageViewController () {
-    int page_number;
+    NSUInteger page_number;
 }
 
 @end
